+++ plone/app/contenttypes/migration/__init__.py
@@ -1,0 +1,9 @@
+import pytz
+
+
+def datetime_fixer(dt, zone):
+    timezone = pytz.timezone(zone)
+    if dt.tzinfo is None:
+        return timezone.localize(dt)
+    else:
+        return timezone.normalize(dt)